requirements.txt
requests
pyspark
python-dotenv
kafka-python
schedule